修改为纯客户端GUI
